refactor(database): created separated files for functions related to files and users
